// src/types/CyberPrice.ts
export type CyberPrice = {
  jobId: string;
  inputValues:string;
  manufacturer?: string;
  productNumber?: string;
  reference?: string;
  searchName: string;
  secondPassword?: string;
  secondaryProductNumber?: string;
  sitePassword?: string;
  siteUsername?: string;
  sites: string;
  vertical?: string;
  zipCode?: string;
};